requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
Pillow==10.1.0
coloredlogs==15.0.1
msedge-selenium-tools==3.141.4